==> 02_DefineMtgPrisoners.Rmd <==
---
title: "02_DefineMtgPrisoners"
author: "Maria Jose Herrera"
date: "6/19/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library("tidyverse")
library("stringr")
library("tibble")

```

```{r}
# Set working directory
setwd("~/Documents/LSE_new/03_Dissertation/Understanding_Society/dissertation")

```

# Definition of mortgage prisoners 

Define mortgage prisoners: 1) self-employed 
                           2) interest-only mortgages
                           3) spending >33% income on mortgage repayment 
                              OR report having had a hard time making mortgage repayments over the last 12 months


```{r}
library("tidyverse")

# Use glob when reading in all data
clean_files <- Sys.glob(file.path("/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/*_clean.rds"), dirmark = FALSE) ## gets the name of all files

# Read in data
# w6_data <- readRDS("w6_clean.rds") # test set
data_files <- lapply(clean_files, readRDS)

# Regex to change the name of dfs 
new_col_names <- str_replace_all(colnames(data_files[[2]]), "^[^_]+(?=_)_", "") # remove all wave prefixes to have uniform col names across waves

# Add mgold as null for wave 1
data_files[[1]] <- add_column(data_files[[1]], mgold = rep(NA, nrow(data_files[[1]])), .after = 30)

# Rename all files
renamed_files <- lapply(data_files[-10], setNames, nm = new_col_names) # rename columns for all except xwave

# Rbind all waves into one df
df_all <- do.call("rbind", renamed_files) # bind all except xwave

# Only consider households w/ a mortgage
df_all <- df_all[!is.na(df_all$mgtype),]

# Add in only necessary xwave data for pidp
df_all <- left_join(x = df_all, y = data_files[[10]], by = "pidp")


```

# Re-categorise vars
## jbstat
```{r}
df_all <- df_all %>%
    mutate(jbstat_recat = case_when(jbstat == 1 ~ 'self-employed',
                                  jbstat == 2 ~ 'employed',
                                  jbstat == 3 ~ 'unemployed',
                                  jbstat == 4 ~ 'not seeking work',
                                  jbstat == 5 ~ 'maternity leave',
                                  jbstat == 6 ~ 'not seeking work',
                                  jbstat == 7 ~ 'not seeking work',
                                  jbstat == 8 ~ 'LT sick/disabled',
                                  jbstat == 9 ~ 'not seeking work',
                                  jbstat == 10 ~ 'employed',
                                  jbstat == 11 ~ 'employed',
                                  TRUE ~ 'Something else'))

# Make reference employed people
df_all$jbstat_recat <- as.factor(df_all$jbstat_recat)
df_all <- within(df_all, jbstat_recat <- relevel(jbstat_recat, ref = 'employed'))


```

## marstat
```{r}
df_all <- df_all %>%
  mutate(marstat_recat = case_when(marstat == 1 ~ 'never married',
                                   marstat == 2 ~ 'married/civil partner',
                                   marstat == 3 ~ 'married/civil partner',
                                   marstat == 4 ~ 'separated',
                                   marstat == 5 ~ 'divorced/dissolved',
                                   marstat == 6 ~ 'widowed/surviving partner',
                                   marstat == 7 ~ 'separated',
                                   marstat == 8 ~ 'divorced/dissolved',
                                   marstat == 9 ~ 'widowed/surviving partner',
                                   TRUE ~ 'something else'))

# Make reference single people
df_all$marstat_recat <- as.factor(df_all$marstat_recat)
df_all <- within(df_all, marstat_recat <- relevel(marstat_recat, ref = 'never married'))



```

## racel_dv
```{r}
df_all <- df_all %>%
  mutate(racel_recat = case_when(racel_dv.x == 1 ~ 'white',
                                   racel_dv.x == 2 ~ 'white',
                                   racel_dv.x == 3 ~ 'white',
                                   racel_dv.x == 4 ~ 'white',
                                   racel_dv.x == 5 ~ 'mixed',
                                   racel_dv.x == 6 ~ 'mixed',
                                   racel_dv.x == 7 ~ 'mixed',
                                   racel_dv.x == 8 ~ 'mixed',
                                   racel_dv.x == 9 ~ 'asian/asian british',
                                   racel_dv.x == 10 ~ 'asian/asian british',
                                   racel_dv.x == 11 ~ 'asian/asian british',
                                   racel_dv.x == 12 ~ 'asian/asian british',
                                   racel_dv.x == 13 ~ 'asian/asian british',
                                   racel_dv.x == 14 ~ 'black/black british',
                                   racel_dv.x == 15 ~ 'black/black british',
                                   racel_dv.x == 16 ~ 'black/black british',
                                   racel_dv.x == 17 ~ 'other ethnic group',
                                   TRUE ~ 'other ethnic group'))

df_all$racel_dv.y <- NULL # drop racel_dv.y since it's the same thing as its .x counterpart

# Make reference white people
df_all$racel_recat <- as.factor(df_all$racel_recat)
df_all <- within(df_all, racel_recat <- relevel(racel_recat, ref = 'white'))


```

## jbstat
```{r}
df_all <- df_all %>%
    mutate(jbstat_recat = case_when(jbstat == 1 ~ 'self-employed',
                                  jbstat == 2 ~ 'employed',
                                  jbstat == 3 ~ 'unemployed',
                                  jbstat == 4 ~ 'not seeking work',
                                  jbstat == 5 ~ 'maternity leave',
                                  jbstat == 6 ~ 'not seeking work',
                                  jbstat == 7 ~ 'not seeking work',
                                  jbstat == 8 ~ 'LT sick/disabled',
                                  jbstat == 9 ~ 'not seeking work',
                                  jbstat == 10 ~ 'employed',
                                  jbstat == 11 ~ 'employed',
                                  TRUE ~ 'Something else'))

# Make reference employed people
df_all$jbstat_recat <- as.factor(df_all$jbstat_recat)
df_all <- within(df_all, jbstat_recat <- relevel(jbstat_recat, ref = 'employed'))


```

## ukborn
```{r}
## ukborn
df_all <- df_all %>%
    mutate(ukborn_recat = case_when(ukborn == 1 ~ 1,
                                  ukborn == 2 ~ 1,
                                  ukborn == 3 ~ 1,
                                  ukborn == 4 ~ 1,
                                  ukborn == 5 ~ 2,
                                  TRUE ~ 0))

# Make reference employed people
df_all$ukborn_recat <- as.factor(df_all$ukborn_recat)
df_all <- within(df_all, ukborn_recat <- relevel(ukborn_recat, ref = 1))

```

# Save df
```{r}
# Save all mortgage owners
saveRDS(df_all, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_allmtgs.rds")

```

# Only look at head of households
```{r}
# look at hrpid variable
df_hoh <- df_all[df_all$pidp == df_all$hrpid, ] # keep only heads of household in data set

```

# Condition 1: Self-employed
```{r}
# Condition 1: Self-employed
# Variable: w_jbstat == 1 (self-employed)

# wave_number <- 6
# w_jbstat <- paste0(letters[wave_number], "_jbstat")
# w6_cond1 <- filter(w6_data, get(w_jbstat) == 1) 

cond1 <- filter(df_hoh, jbstat == 1) # need get() since value of variable is the actual column name
saveRDS(cond1, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_cond1.rds")

```

# Condition 2: Interest only mortgage
```{r}
# Condition 2: Interest only mortgage
# Variable: w_mgtype == 4 (interest-only)

# wave_number <- 6
# w_mgtype <- paste0(letters[wave_number], "_mgtype")
# w6_cond2 <- filter(w6_data, get(w_mgtype) == 4)

cond2 <- filter(df_hoh, mgtype == 4) # need get() since value of variable is the actual column name
saveRDS(cond2, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_cond2.rds")

```

# Condition 3: Affordability
```{r}
# Condition 3: spending >33% income on mortgage repayment OR report having had a hard time making mortgage repayments over the last 12 months

# Variable 1: fihhmnnet1_dv (total monthly income, no deductions)

# wave_number <- 6
# w_fihhmnnet1_dve <- paste0(letters[wave_number], "_fihhmnnet1_dv")

# Variable 2: houscost1_dv (monthly housing cost /in/cluding mortgage principal payments) 

# wave_number <- 6
# w_houscost1_dv <- paste0(letters[wave_number], "_houscost1_dv")


# Variable 3: houscost2_dv (monthly housing cost /ex/cluding mortgage principal payments)

# wave_number <- 6
# w_houscost2_dv <- paste0(letters[wave_number], "_houscost2_dv")


# (Var 2 - Var 3) / Var 1 = % income spent on mtg repayment
cond3_1 <- df_hoh %>% 
  mutate(
    mtg_principal  = houscost1_dv - houscost2_dv,
    pct_monthly = mtg_principal / fihhmnnet1_dv) %>%
  filter(mtg_principal>0, pct_monthly > 0.33, pct_monthly != Inf)

cond3_1 <- subset(cond3_1, select = -c(mtg_principal, pct_monthly)) # dropping columns to be able to merge df's

saveRDS(cond3_1, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/cond3_1.rds")


## TODO: graph and see what's going on here

# Variable 4: w_xphsdb == 1 (reported difficulty paying for housing)

# wave_number <- 6
# w_xphsdb <- paste0(letters[wave_number], "_xphsdb")

# cond3_2 <- filter(df_hoh, xphsdb == 1) # need get() since value of variable is the actual column name
old_df <- readRDS("/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/cond3_2.rds")

cond32_ids <- old_df$pidp

cond3_2 <- df_hoh[df_hoh$pidp %in% cond32_ids, ]

saveRDS(cond3_2, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/updtd_cond3_2.rds")


```

# Combine into 1 df

```{r}
# Put all df's in a list
ls_dfs <- list(cond1, cond2, cond3_1, cond3_2)

# Do call to bind all df's in the list
df_allcond <- do.call("rbind", ls_dfs)

```


# Get rid of duplicates (i.e. keeping any rows that meet any of the 1 above conditions)

```{r}
# Set variable that we want distinct
unique_mtgprisoners <- distinct(df_allcond, pidp, .keep_all = TRUE)
unique_mtgprisoners$is_mtgprisoner <- 1

# TODO: get rid of incomplete observations?
# TODO: record that we have 28,259 mortgage prisoners! ACTUALLY 4474 UNIQUE MTG PRISONERS IN UPDATED VERSION
# TODO: explain how i got to this number in writing!
```

# Save df of just mtg prisoners head of households
```{r}
# Save df of mtg prisoners
saveRDS(unique_mtgprisoners, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_mtgprisoners_hoh.rds")

```

# Get all of the people in the household of mtg prisoners
```{r}
df_allmtgprisoners_inhh <- df_all[df_all$hrpid %in% unique_mtgprisoners$pidp, ]

# validity check - the number of unique heads of household is the same as the nrow() of unique_mtgprisoners
length(unique(df_allmtgprisoners_inhh$hrpid)) ## 4474
nrow(unique_mtgprisoners) ## 4474


```

# Now add column to df_all that indicates mtg prisoner status
```{r}
# Get id's of mtg prisoners
mtgprisoner_ids <- paste0(df_allmtgprisoners_inhh$pidp, df_allmtgprisoners_inhh$hidp) # trying to get same number (28k) as before since PIDP+ HIDP is unique

# Add column, where == 1 if mtg prisoner and == 0 if not
df_all$is_mtgprisoner <- ifelse(paste0(df_all$pidp, df_all$hidp) %in% mtgprisoner_ids, 1, 0)

# Save df
saveRDS(df_all, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_all_prisonerstatus.rds")
```

# All people before and after law
```{r}
# All prelegis
df_prelegis <- df_all[df_all$wavenumber < 6, ]

# All post legis
df_postlegis <- df_all[df_all$wavenumber >= 6, ]

# Save df's
saveRDS(df_prelegis, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_all_prelegis.rds")
saveRDS(df_postlegis, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_all_postlegis.rds")


```

# Mtg prisoners after 2014 (wave 6 - wave 9)
```{r}
# Create mtg prisoners df from when legislation was passed (2014) onwards
df_mtgprisoners_postlegis <- df_allmtgprisoners_inhh[df_allmtgprisoners_inhh$wavenumber >= 6, ]

# Save df
saveRDS(df_mtgprisoners_postlegis, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_mtgprisoners_postlegis.rds")

```

# Mtg prisoners before 2014 (wave 1 - wave 6)
```{r}
# Create mtg prisoners df from when legislation was passed (2014) onwards
df_mtgprisoners_prelegis <- df_allmtgprisoners_inhh[df_allmtgprisoners_inhh$wavenumber < 6, ]

# Save df
saveRDS(df_mtgprisoners_prelegis, "/Users/mariajoseherrera/Documents/LSE_new/03_Dissertation/Understanding_Society/df_mtgprisoners_prelegis.rds")

```
